PLAT-1260 Improve testability, add tests
This includes a major refactoring of connection, which reverts somewhat to its original
form, but in a cleaner fashion.

It changes a few key elements in the triggering so that dependencies can be passed as
parameters, which improves testability. Some parameters were moved from constructor
to methods, making it possible to inject default implementations for these dependencies.

The test suite has been enhanced with further testing of trigger rules.

The major components left untested are UI, settings and settings service, and the
HTTP request itself.