PLAT-1260 Fix settings branch retrieval
This is a perfect example of code coverage not meaning lack of defects.

The code in question was fully tested with multiple cases both on the
caller and callee sides. However, the meaning of one parameter was
inverted on each side.

This kind of error can be prevented by ensuring that, whenever you mock
something, the tests for that something include the sequence of calls
that were mocked. Which was not done here (and still isn't -- I'm trying
to get a fix out fast).